refactor: move ThematicSearch and merge hooks
