+++ src/features/search/pages/ThematicSearch/Films/FilmsPlaylist.native.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
-import * as useFilmsOffersAPI from 'features/search/pages/Search/ThematicSearch/Films/algolia/useFilmsOffers'
-import { FilmsPlaylist } from 'features/search/pages/Search/ThematicSearch/Films/FilmsPlaylist'
-import { filmsPlaylistAlgoliaSnapshot } from 'features/search/pages/Search/ThematicSearch/Films/fixtures/filmsPlaylistAlgoliaSnapshot'
+import * as useThematicSearchPlaylistsAPI from 'features/search/pages/ThematicSearch/api/useThematicSearchPlaylists'
+import { FilmsPlaylist } from 'features/search/pages/ThematicSearch/Films/FilmsPlaylist'
+import { filmsPlaylistAlgoliaSnapshot } from 'features/search/pages/ThematicSearch/Films/fixtures/filmsPlaylistAlgoliaSnapshot'
 import * as useFeatureFlagAPI from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { LocationMode, Position } from 'libs/location/types'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
@@ -26,10 +26,12 @@ jest.mock('libs/location/LocationWrapper', () => ({
 
 const DEFAULT_PLAYLIST_TITLE = 'Vidéos et documentaires'
 const DEFAULT_PLAYLIST_OFFERS = filmsPlaylistAlgoliaSnapshot
-const useFilmsOffersSpy = jest.spyOn(useFilmsOffersAPI, 'useFilmsOffers').mockReturnValue({
-  playlists: DEFAULT_PLAYLIST_OFFERS,
-  isLoading: false,
-})
+const useThematicSearchPlaylistsSpy = jest
+  .spyOn(useThematicSearchPlaylistsAPI, 'useThematicSearchPlaylists')
+  .mockReturnValue({
+    playlists: DEFAULT_PLAYLIST_OFFERS,
+    isLoading: false,
+  })
 
 describe('Films', () => {
   it('should render playlist when algolia returns offers', async () => {
@@ -39,7 +41,7 @@ describe('Films', () => {
   })
 
   it('should not render playlist when algolia does not return offers', async () => {
-    useFilmsOffersSpy.mockReturnValueOnce({ playlists: [], isLoading: false })
+    useThematicSearchPlaylistsSpy.mockReturnValueOnce({ playlists: [], isLoading: false })
     renderCinema()
 
     expect(screen.queryByText(DEFAULT_PLAYLIST_TITLE)).not.toBeOnTheScreen()
