+++ src/features/search/pages/ThematicSearch/ThematicSearch.web.test.tsx
@@ -5,7 +5,7 @@ import { SearchGroupNameEnumv2, SubcategoriesResponseModelv2 } from 'api/gen'
 import { contentfulGtlPlaylistSnap } from 'features/gtlPlaylist/fixtures/contentfulGtlPlaylistSnap'
 import { initialSearchState } from 'features/search/context/reducer'
 import * as useSearch from 'features/search/context/SearchWrapper'
-import { ThematicSearch } from 'features/search/pages/Search/ThematicSearch/ThematicSearch'
+import { ThematicSearch } from 'features/search/pages/ThematicSearch/ThematicSearch'
 import { env } from 'libs/environment'
 import { LocationMode } from 'libs/location/types'
 import { PLACEHOLDER_DATA } from 'libs/subcategories/placeholderData'
