+++ src/features/search/pages/ThematicSearch/ThematicSearchBar.tsx
