+++ src/features/search/pages/ThematicSearch/ThematicSearchPlaylist.tsx
@@ -2,7 +2,7 @@ import { Hit } from '@algolia/client-search'
 import React from 'react'
 
 import { Referrals, ScreenNames } from 'features/navigation/RootNavigator/types'
-import { ThematicSearchPlaylistData } from 'features/search/pages/Search/ThematicSearch/types'
+import { ThematicSearchPlaylistData } from 'features/search/pages/ThematicSearch/types'
 import { useTransformOfferHits } from 'libs/algolia/fetchAlgolia/transformOfferHit'
 import { Offer } from 'shared/offer/types'
 import { useRenderPassPlaylist } from 'shared/renderPassPlaylist'
@@ -11,7 +11,7 @@ import { LENGTH_M, RATIO_HOME_IMAGE } from 'ui/theme'
 
 const PLAYLIST_ITEM_HEIGHT = LENGTH_M
 const PLAYLIST_ITEM_WIDTH = LENGTH_M * RATIO_HOME_IMAGE
-export interface CinemaPlaylistProps {
+export interface ThematicSearchPlaylist {
   playlist: ThematicSearchPlaylistData
   analyticsFrom: Referrals
   route: Extract<ScreenNames, 'ThematicSearch'>
@@ -21,7 +21,7 @@ export function ThematicSearchPlaylist({
   playlist,
   analyticsFrom,
   route,
-}: Readonly<CinemaPlaylistProps>) {
+}: Readonly<ThematicSearchPlaylist>) {
   const renderPassPlaylist = useRenderPassPlaylist({ analyticsFrom, route, playlist })
   const transformOfferHits = useTransformOfferHits()
   return (
