+++ src/features/search/pages/ThematicSearch/types.ts
@@ -1,11 +1,6 @@
 import { AlgoliaHit } from 'libs/algolia/types'
-import { Position } from 'libs/location'
 
 export type ThematicSearchPlaylistData = {
   title: string
   offers: { hits: AlgoliaHit[] }
-}
-
-export type FetchThematicSearchOffers = {
-  userLocation?: Position
 }
